add threads to chip8 framework

+++ chip8/src/input.h
@@ -1,1 +1,3 @@
-int process_input();
+#include <stdint.h>
+
+int8_t process_input();
